# src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"IUserData": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"userName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"userName",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserDataWithoutUserId": {
				"properties": {
					"userName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"userName",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserDataLogin": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserDataWithoutPassword": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"userName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"userName",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "ntt-music-server",
		"version": "1.0.0",
		"description": "API Ntt Music",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/user/sign-up": {
			"post": {
				"summary": "Cria um novo usuário",
				"operationId": "AddUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserData"
								}
							}
						}
					}
				},
				"tags": [
					"Acesso de Usuário"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserDataWithoutUserId"
							}
						}
					}
				}
			}
		},
		"/user/login": {
			"post": {
				"summary": "Realiza o login do usuário",
				"operationId": "LoginUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Acesso de Usuário"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserDataLogin"
							}
						}
					}
				}
			}
		},
		"/user": {
			"get": {
				"summary": "Lista todos os usuários",
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IUserDataWithoutPassword"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Acesso de Usuário"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/user/{userId}": {
			"get": {
				"summary": "Busca um usuário pelo ID",
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserDataWithoutPassword"
								}
							}
						}
					}
				},
				"tags": [
					"Acesso de Usuário"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}